+++ front/app/api/auth/[...nextauth]/route.ts
@@ -1,44 +1,49 @@
 import NextAuth from 'next-auth';
-import GoogleProvider from 'next-auth/providers/google';
-import axios from 'axios';
+// import GoogleProvider from 'next-auth/providers/google';
+// import axios from 'axios';
+import { nextAuthOptions } from '@/lib/next-auth/options';
 
-const apiUrl = process.env.NEXT_PUBLIC_API_URL;
+// const apiUrl = process.env.NEXT_PUBLIC_API_URL;
 
-const handler = NextAuth({
-	secret: process.env.NEXTAUTH_SECRET,
-	providers: [
-		GoogleProvider({
-			clientId: process.env.GOOGLE_CLIENT_ID ?? '',
-			clientSecret: process.env.GOOGLE_CLIENT_SECRET ?? '',
-		}),
-	],
-	callbacks: {
-		async signIn({ user, account }) {
-			const provider = account?.provider;
-			const uid = user?.id;
-			const name = user?.name;
-			const email = user?.email;
-			try {
-				const response = await axios.post(
-					`${apiUrl}/auth/${provider}/callback`,
-					{
-						provider,
-						uid,
-						name,
-						email,
-					}
-				);
-				if (response.status === 200) {
-					return true;
-				} else {
-					return false;
-				}
-			} catch (error) {
-				console.log('エラー', error);
-				return false;
-			}
-		},
-	},
-});
+
+const handler = NextAuth(nextAuthOptions);
+
+
+// const handler = NextAuth({
+// 	secret: process.env.NEXTAUTH_SECRET,
+// 	providers: [
+// 		GoogleProvider({
+// 			clientId: process.env.GOOGLE_CLIENT_ID ?? '',
+// 			clientSecret: process.env.GOOGLE_CLIENT_SECRET ?? '',
+// 		}),
+// 	],
+// 	callbacks: {
+// 		async signIn({ user, account }) {
+// 			const provider = account?.provider;
+// 			const uid = user?.id;
+// 			const name = user?.name;
+// 			const email = user?.email;
+// 			try {
+// 				const response = await axios.post(
+// 					`${apiUrl}/auth/${provider}/callback`,
+// 					{
+// 						provider,
+// 						uid,
+// 						name,
+// 						email,
+// 					}
+// 				);
+// 				if (response.status === 200) {
+// 					return true;
+// 				} else {
+// 					return false;
+// 				}
+// 			} catch (error) {
+// 				console.log('エラー', error);
+// 				return false;
+// 			}
+// 		},
+// 	},
+// });
 export { handler as GET, handler as POST };
