+++ front/lib/next-auth/options.ts
@@ -1,0 +1,52 @@
+import GoogleProvider from 'next-auth/providers/google';
+import type { NextAuthOptions, Session } from "next-auth";
+import axios from 'axios';
+
+const apiUrl = process.env.NEXT_PUBLIC_API_URL;
+
+export const nextAuthOptions: NextAuthOptions = {
+	secret: process.env.NEXTAUTH_SECRET,
+	providers: [
+		GoogleProvider({
+			clientId: process.env.GOOGLE_CLIENT_ID ?? '',
+			clientSecret: process.env.GOOGLE_CLIENT_SECRET ?? '',
+		}),
+	],
+	callbacks: {
+		session: ({ session }:any) => { // 4⃣
+      return {
+        ...session,
+        user: {
+          ...session.user
+        },
+      };
+    },
+		async signIn({ user, account }) {
+			const provider = account?.provider;
+			const uid = user?.id;
+			const name = user?.name;
+			const email = user?.email;
+			try {
+				const response = await axios.post(
+					`${apiUrl}/auth/${provider}/callback`,
+					{
+						provider,
+						uid,
+						name,
+						email,
+					}
+				);
+				if (response.status === 200) {
+					return true;
+				} else {
+					return false;
+				}
+			} catch (error) {
+				console.log('エラー', error);
+				return false;
+			}
+		},
+	},
+}
+
+
